--- tagstudio/src/qt/helpers/__init__.py
@@ -1,3 +1,0 @@
-from .file_opener import open_file, FileOpenerHelper, FileOpenerLabel
-from .function_iterator import FunctionIterator
-from .custom_runnable import CustomRunnable
