+++ tagstudio/src/qt/modals/fix_unlinked.py
@@ -10,9 +10,11 @@ from PySide6.QtCore import QThread, Qt, QThreadPool
 from PySide6.QtWidgets import QWidget, QVBoxLayout, QHBoxLayout, QLabel, QPushButton
 
 from src.core.library import Library
-from src.qt.helpers import FunctionIterator, CustomRunnable
-from src.qt.modals import DeleteUnlinkedEntriesModal, RelinkUnlinkedEntries
-from src.qt.widgets import ProgressWidget
+from src.qt.helpers.function_iterator import FunctionIterator
+from src.qt.helpers.custom_runnable import CustomRunnable
+from src.qt.modals.delete_unlinked import DeleteUnlinkedEntriesModal
+from src.qt.modals.relink_unlinked import RelinkUnlinkedEntries
+from src.qt.widgets.progress import ProgressWidget
 
 # Only import for type checking/autocompletion, will not be imported at runtime.
 if typing.TYPE_CHECKING:
