+++ tagstudio/src/qt/modals/relink_unlinked.py
@@ -7,8 +7,9 @@ import typing
 from PySide6.QtCore import QObject, Signal, QThreadPool
 
 from src.core.library import Library
-from src.qt.helpers import FunctionIterator, CustomRunnable
-from src.qt.widgets import ProgressWidget
+from src.qt.helpers.function_iterator import FunctionIterator
+from src.qt.helpers.custom_runnable import CustomRunnable
+from src.qt.widgets.progress import ProgressWidget
 
 # Only import for type checking/autocompletion, will not be imported at runtime.
 if typing.TYPE_CHECKING:
