+++ tagstudio/src/qt/modals/tag_search.py
@@ -19,7 +19,8 @@ from PySide6.QtWidgets import (
 
 from src.core.library import Library
 from src.core.palette import ColorType, get_tag_color
-from src.qt.widgets import PanelWidget, TagWidget
+from src.qt.widgets.panel import PanelWidget
+from src.qt.widgets.tag import TagWidget
 
 
 ERROR = f"[ERROR]"
