--- tagstudio/src/qt/widgets/__init__.py
@@ -1,13 +1,0 @@
-from .fields import FieldContainer, FieldWidget
-from .collage_icon import CollageIconRenderer
-from .thumb_button import ThumbButton
-from .thumb_renderer import ThumbRenderer
-from .panel import PanelWidget, PanelModal
-from .text_box_edit import EditTextBox
-from .text_line_edit import EditTextLine
-from .progress import ProgressWidget
-from .tag import TagWidget
-from .tag_box import TagBoxWidget
-from .text import TextWidget
-from .item_thumb import ItemThumb
-from .preview_panel import PreviewPanel
